Respawn all players on round end (incomplete)
When a player's health hits 0 or less, a HUD is set that displays
VICTORY or DEFEAT, and a respawn timer starts.

I'm close, but I'm having trouble figuring out how to get a respawn
timer shown on the clients while still letting the server remain in
control or respawning. Almost there, I'll continue to work on it after
some sleep.