Using HUDState RepNotify to change HUDs
Rather than having the server explicit call/use the Set Client HUD node,
I've decided to set the HUDState enum variable in the MyPlayerController
BP to be RepNotify variable. If the HUDState changes, the UI/HUD for the
player will change accordingly.

It's not working as it should, unfortunately, but I suspect it's due to
the fact that I create and add widgets to the viewport willynilly
without properly removing or reusing existing widgets.