Set Client HUD working, scoreboard implementing
Set Client HUD is now working as expected - the trick was to make use of
the event "Destroyed" in the HUD class, and make it call "Remove from
parent" on all the widgets. This cleans everything up for the next
HUD/set of widgets.

I've also begun implementing the Score widget - the widget itself seems
to be functionally fine, pulling player scores from MyGameState, but I'm
running into errors when trying to remove the score widget. "Remove from
parent" won't work on it. I'll have to keep at it.