Round counter/number
Added a RoundNumber integer to MyGameState so that a goal score can be
set that will officially end a match. Also implemented a widget that
appears at the top of the players' screens that keeps track of what the
current round is.

Up next is deciding and implementing what should happen when a match
point occurs.