Renamed assets, prepping for main menu and lobby
I've renamed assets so that they're more consistent, and identifiable
(e.g. prefacing HUD in front of all HUD-derived classes). I've also
decided to move forwarded with implementing the main menu and lobby
functionality. It should be fairly straightforward, as I plan to just
copy the implementation in the MultiplayerShootout example game by Epic.
Game end conditions and events are on the backburner for now, since I
want some sort of "Play Again?" option to appear, and that needs to have
the menu and lobby implemented (and the main menu stuff sounds more
exciting right now, ha!).